requirements.txt
amazon_transcribe==0.6.2
numpy==1.26.4
openai==1.26.0
opencv_python==4.9.0.80
Pillow==10.3.0
PyDrive==1.3.1
pyvirtualcam==0.11.1
Requests==2.31.0
sounddevice==0.4.6
